Split the Resources class for file specific stuff
